All Client Side API Endpoints are successfully Connected to MongoDB
Accepted Endpoints
- /auth/signup [POST]
- /auth/login [POST]
- /auth/set-pin [POST]
- /wallet/balance [GET]
- /transaction/transfer [POST]
- /transaction/history [GET]
- /profile [GET]

+++ src/main/java/com/tksolutions/astraguard/dto/TransactionTransferRequest.java
@@ -6,12 +6,11 @@ import lombok.Data;
 public class TransactionTransferRequest {
 
     private String toUpi;
-    private Double amount;
+    private Long amount;
     private String pin;
     private String transactionType;
     private Device device;
     private Location location;
-
 
 
     @Data
